+++ AndroidSDK/src/com/leanplum/LeanplumNotificationHelper.java
@@ -21,11 +21,18 @@
 package com.leanplum;
 
 import android.app.Notification;
+import android.app.PendingIntent;
 import android.content.Context;
+import android.content.res.Resources;
+import android.graphics.Bitmap;
+import android.os.Build;
 import android.os.Bundle;
 import android.support.v4.app.NotificationCompat;
 import android.text.TextUtils;
+import android.util.TypedValue;
+import android.widget.RemoteViews;
 
+import com.leanplum.internal.Constants;
 import com.leanplum.internal.JsonConverter;
 import com.leanplum.internal.Log;
 import com.leanplum.utils.BuildUtil;
@@ -38,6 +45,10 @@ import java.util.Map;
  * @author Anna Orlova
  */
 class LeanplumNotificationHelper {
+
+  private static final int BIGPICTURE_TEXT_TOP_PEDDING = -14;
+  private static final int BIGPICTURE_TEXT_SIZE = 14;
+
   /**
    * If notification channels are supported this method will try to create
    * NotificationCompat.Builder with default notification channel if default channel id is provided.
@@ -105,7 +116,7 @@ class LeanplumNotificationHelper {
   @SuppressWarnings("deprecation")
   static NotificationCompat.Builder getNotificationCompatBuilder(Context context, Bundle message) {
     NotificationCompat.Builder builder = null;
-    // If we are targeting API 26, try to find supplied chLannel to post notification.
+    // If we are targeting API 26, try to find supplied channel to post notification.
     if (BuildUtil.isNotificationChannelSupported(context)) {
       try {
         String channel = message.getString("lp_channel");
@@ -170,5 +181,72 @@ class LeanplumNotificationHelper {
       builder = new Notification.Builder(context);
     }
     return builder;
+  }
+
+  /**
+   * Gets Notification.Builder with 2 lines at BigPictureStyle notification text.
+   *
+   * @param context The application context.
+   * @param message Push notification Bundle.
+   * @param contentIntent PendingIntent.
+   * @param title String with title for push notification.
+   * @param messageText String with text for push notification.
+   * @param bigPicture Bitmap for BigPictureStyle notification.
+   * @return Notification.Builder or null.
+   */
+  static Notification.Builder getNotificationBuilder(Context context, Bundle message,
+      PendingIntent contentIntent, String title, final String messageText, Bitmap bigPicture) {
+    if (Build.VERSION.SDK_INT < 16) {
+      return null;
+    }
+    Notification.Builder notificationBuilder =
+        getNotificationBuilder(context, message);
+    notificationBuilder.setSmallIcon(context.getApplicationInfo().icon)
+        .setContentTitle(title)
+        .setContentText(messageText);
+    Notification.BigPictureStyle bigPictureStyle = new Notification.BigPictureStyle() {
+      @Override
+      protected RemoteViews getStandardView(int layoutId) {
+        RemoteViews remoteViews = super.getStandardView(layoutId);
+        // Modifications of stanxdard push RemoteView.
+        try {
+          int id = Resources.getSystem().getIdentifier("text", "id", "android");
+          remoteViews.setBoolean(id, "setSingleLine", false);
+          remoteViews.setInt(id, "setLines", 2);
+          if (Build.VERSION.SDK_INT < 23) {
+            // Make text smaller.
+            remoteViews.setViewPadding(id, 0, BIGPICTURE_TEXT_TOP_PEDDING, 0, 0);
+            remoteViews.setTextViewTextSize(id, TypedValue.COMPLEX_UNIT_SP, BIGPICTURE_TEXT_SIZE);
+          }
+        } catch (Throwable throwable) {
+          Log.e("Cannot modify push notification layout.");
+        }
+        return remoteViews;
+      }
+    };
+
+    bigPictureStyle.bigPicture(bigPicture)
+        .setBigContentTitle(title)
+        .setSummaryText(message.getString(Constants.Keys.PUSH_MESSAGE_TEXT));
+    notificationBuilder.setStyle(bigPictureStyle);
+
+    if (Build.VERSION.SDK_INT >= 24) {
+      // By default we cannot reach getStandardView method on API>=24. I we call
+      // createBigContentView, Android will call getStandardView method and we can get
+      // modified RemoteView.
+      try {
+        RemoteViews remoteView = notificationBuilder.createBigContentView();
+        if (remoteView != null) {
+          // We need to set received RemoteView as a custom big content view.
+          notificationBuilder.setCustomBigContentView(remoteView);
+        }
+      } catch (Throwable t) {
+        Log.e("Cannot modify push notification layout.", t);
+      }
+    }
+
+    notificationBuilder.setAutoCancel(true);
+    notificationBuilder.setContentIntent(contentIntent);
+    return notificationBuilder;
   }
 }
